+++ ecohack-api/src/main/java/de/dlh/lhind/ecohack/model/entity/FoodProvider.java
@@ -39,10 +39,10 @@ public class FoodProvider{
     private String nipt;
 
     @Column
-    private Long latitude;
+    private Double latitude;
 
     @Column
-    private Long longtitude;
+    private Double longtitude;
 
 
 }
